extract user service tests
